Refactor controllers/ directory out of server/db/ and into server/
